app/backend/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
reportlab
beautifulsoup4
google-generativeai
python-dotenv